Fix issue with react not rerendering game Add .clone method to Game class
Cloning when updating state tells react the object is different
so will ensure it rerenders. The issue was that even though the
object was being mutated, it's still the same object, and react
doesn't consider that enough to rerender.

Using reducers is another solution to this, but that's complicated
for where we are.